images are now removed from filesystem when user deletes a product

+++ controllers/admin.js
@@ -1,8 +1,8 @@
 const mongoose = require('mongoose');
-
 const {validationResult} = require('express-validator');
 
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   if (!req.session.isLoggedIn) {
@@ -158,6 +158,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = req.body.description;
       //if there is an image, save the path to the DB
       if (req.file) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path
       }
       return product.save()
@@ -191,7 +192,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({_id: prodId, userId: req.user._id});
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
